requirements.txt
Flask==2.1.1
google-cloud-vision==2.7.0
